Refactor code in updating user follow

+++ main/libs/firebase/user.py
@@ -1,6 +1,6 @@
 from main import errors
 from werkzeug.security import generate_password_hash, check_password_hash
-from main.libs.firebase.image import users_ref
+from main.libs.firebase import users_ref
 
 
 def create_user(email=None, password=None):
@@ -23,7 +23,7 @@ def check_user_credentials(email=None, password=None):
         user_key = get_user_key_or_none(email)
         if not user_key:
             return False
-        if not password:
+        if password:
             hashed_password_in_db = users_ref.child(user_key).get()['password']
             if not check_password_hash(pwhash=hashed_password_in_db, password=password):
                 return False
@@ -63,24 +63,32 @@ def update_user_follow(follower_email, followee_email):
         followee_key = get_user_key_or_none(email=followee_email)
         if not follower_key or not followee_key:
             raise errors.UserDoesNotExist()
-        followee_followers = users_ref.child(followee_key).child('followers').get()
-        followed = False
 
-        for key, value in followee_followers:
-            if value['email'] == follower_email:
-                followed = True
-                users_ref.child(followee_key).child('followers').child(key).delete()
-            if not followed:
-                users_ref.child(followee_key).child('followers').push({'email': follower_email})
+        followee_followers = users_ref.child(followee_key).child('followers').get()
         follower_followees = users_ref.child(follower_key).child('followees').get()
 
-        for key, value in follower_followees:
-            if value['email'] == followee_email:
-                if followed:
-                    users_ref.child(follower_key).child('followees').child(key).delete()
-                else:
-                    users_ref.child(follower_key).child('followees').push({'email': followee_email})
+        followed = False
+        if followee_followers:
+            for key, value in followee_followers.items():
+                if value['email'] == follower_email:
+                    followed = True
+                    users_ref.child(followee_key).child('followers').child(key).delete()
 
-        return not followed
+            if not followed:
+                users_ref.child(followee_key).child('followers').push({'email': follower_email})
+        else:
+            users_ref.child(followee_key).child('followers').push({'email': follower_email})
+
+        if follower_followees:
+            for key, value in follower_followees.items():
+                if value['email'] == followee_email:
+                    users_ref.child(follower_key).child('followees').child(key).delete()
+
+            if not followed:
+                users_ref.child(follower_key).child('followees').push({'email': followee_email})
+        else:
+            users_ref.child(follower_key).child('followees').push({'email': followee_email})
+
+        return followed
     except Exception:
         raise Exception('There is an error following another user')
